add: userController and categoryController

+++ ewm-main-service/src/main/java/ru/practicum/ewmmainservice/category/service/CategoryServiceImpl.java
@@ -1,5 +1,7 @@
 package ru.practicum.ewmmainservice.category.service;
 
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.ewmmainservice.category.dto.CategoryDto;
 import ru.practicum.ewmmainservice.category.dto.NewCategoryDto;
@@ -7,7 +9,9 @@ import ru.practicum.ewmmainservice.category.mapper.CategoryMapper;
 import ru.practicum.ewmmainservice.category.repository.CategoryRepository;
 
 import javax.transaction.Transactional;
+import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
@@ -28,18 +32,35 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void removeById(Long id) {
-        categoryRepository.deleteById(id);
+        if(categoryRepository.existsById(id))
+            categoryRepository.deleteById(id);
     }
 
     @Override
     @Transactional
     public CategoryDto update(Long id, CategoryDto dto) {
         var categoryToUpdateOptional = categoryRepository.findById(id);
-        var categoryToUpdate = categoryToUpdateOptional.get();
+        var categoryToUpdate = categoryToUpdateOptional.orElseThrow();
         var name = dto.getName();
         if (name != null && name.length() > 0 && !Objects.equals(name, categoryToUpdate.getName())) {
             categoryToUpdate.setName(name);
         }
         return categoryMapper.toDto(categoryToUpdate);
+    }
+
+    @Override
+    public List<CategoryDto> getAll(Integer from, Integer size) {
+        return categoryRepository.findAll(PageRequest.of(from, size))
+                .map(categoryMapper::toDto)
+                .stream()
+                .collect(Collectors.toUnmodifiableList());
+    }
+
+    @Override
+    public CategoryDto getById(long id) {
+        return categoryRepository
+                .findById(id)
+                .map(categoryMapper::toDto)
+                .orElseThrow();
     }
 }
